refactor: PreferredStyle Enum 위치를 Dto에서 엔티티로 이동
