+++ src/main/java/server/wekick/member/entity/PreferredStyle.java
@@ -1,4 +1,4 @@
-package server.wekick.member.dto;
+package server.wekick.member.entity;
 
 public enum PreferredStyle {
     ATTACK, DEFENSE, GOALKEEPER
